backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
langchain==0.0.340
langchain-community==0.0.10
openai==0.28.0
chromadb==0.4.22
selenium==3.141.0
beautifulsoup4==4.12.2
redis==5.0.1
clickhouse-driver==0.2.6
httpx==0.25.1
pydantic[email]==2.5.0
numpy==1.26.4
webdriver_manager==4.0.2
lxml==4.9.3
urllib3==1.26.18
redisvl==0.13.2